+++ .github/workflows/Release.yaml
@@ -6,12 +6,11 @@ on:
     branches: [ main ]
 
 jobs:
-  release:
-    if: github.event.pull_request.merged == true
+  Release:
+    if: ${{ github.event.pull_request.merged == true }}
     uses: gesslar/Maint/.github/workflows/ReleaseMudlet.yaml@main
+    secrets: inherit
     with:
       quality_check: ""
-      inject_mupdate: false
-    secrets: inherit
     permissions:
       contents: write
