+++ .gitignore
@@ -3,7 +3,6 @@
 build/
 .output
 
-# Ignore any file starting with "muddle" in the project root directory
-# This utility file is specific to individual developers and should not be
-# included in the repo
-/muddle*
+.vscode/
+work/
+node_modules/
