--- .vscode/settings.json
@@ -1,11 +1,0 @@
-{
-  "Lua.diagnostics.globals": [
-    "doSpeedWalk",
-    "mudlet",
-    "gmod"
-  ],
-  "Lua.runtime.version": "Lua 5.1",
-  "Lua.workspace.library": [
-    "/home/gesslar/.config/Code/User/globalStorage/delwing.mudlet-scripts-sdk\\docs"
-  ],
-}
