+++ mfile
@@ -1,10 +1,10 @@
 {
   "package": "OxidusUI",
   "title": "Basic UI for Oxidus mudlib",
-  "description": "Get some visuality with your textuality.",
-  "version": "1.7.1",
+  "description": "",
+  "version": "1.8.0",
   "author": "gesslar",
-  "icon": "",
+  "icon": "src/resources/oxidus.png",
   "dependencies": "",
   "outputFile": true
 }
