+++ package.json
@@ -1,14 +1,17 @@
 {
   "name": "oxidus-ui",
-  "version": "1.7.1",
+  "version": "1.8.0",
   "description": "A Mudlet GUI for the Oxidus mudlib.",
   "scripts": {
-    "build": "npx @gesslar/muddy",
-    "watch": "npx @gesslar/muddy -w",
-    "major": "npx @gesslar/muddy version major",
-    "minor": "npx @gesslar/muddy version minor",
-    "patch": "npx @gesslar/muddy version patch",
+    "build": "npx muddy",
+    "watch": "npx muddy -w",
+    "major": "npx muddy version major --package",
+    "minor": "npx muddy version minor --package",
+    "patch": "npx muddy version patch --package",
     "pr": "gt submit -p --ai"
   },
-  "license": "0BSD"
+  "license": "0BSD",
+  "devDependencies": {
+    "@gesslar/muddy": "^3.2.2"
+  }
 }
