+++ src/scripts/PanelWindow.lua
@@ -1,328 +1,8 @@
----@diagnostic disable-next-line: param-type-mismatch
--- local Splitter = require(f"__PKGNAME__/GeyserSplitter")
-
-function __PKGNAME__.buildUi()
+function __PKGNAME__.BuildPanelWindow()
   -- This table is an array of items in the room
-  __PKGNAME__.inventory.room = __PKGNAME__.inventory.room or {}
+  __PKGNAME__.RoomInventoryList = __PKGNAME__.RoomInventoryList or {}
   -- This table is an array of items in the inventory
-  __PKGNAME__.inventory.inv = __PKGNAME__.inventory.inv or {}
-
-  __PKGNAME__.MainContainer = __PKGNAME__.MainContainer or Geyser.Label:new({
-    name = "MainContainer",
-    x = 0,
-    y = -(__PKGNAME__.metrics.height),
-    width = "100%",
-    height = __PKGNAME__.metrics.height,
-    stylesheet = __PKGNAME__.styles.MainBG,
-  })
-
-  __PKGNAME__.BarBox = __PKGNAME__.BarBox or Geyser.HBox:new({
-    name = "BarBox",
-    x = 0,
-    y = 0,
-    height = "100%",
-    width = "100%",
-  }, __PKGNAME__.MainContainer)
-
-  -- Char.Vitals information
-  __PKGNAME__.VitalsBox = __PKGNAME__.VitalsBox or Geyser.VBox:new({
-    name = "VitalsBox",
-  }, __PKGNAME__.BarBox)
-
-  -- HP
-  __PKGNAME__.HPContainer = __PKGNAME__.HPContainer or Geyser.HBox:new({
-    name = "HPContainer",
-  }, __PKGNAME__.VitalsBox);
-
-  __PKGNAME__.HPLabel = __PKGNAME__.HPLabel or Geyser.Label:new({
-    name = "HPLabel",
-    x = 0,
-    y = 0,
-    height = "100%",
-    width = 30,
-    message = "HP",
-    stylesheet = __PKGNAME__.styles.Label,
-    fontSize = __PKGNAME__.metrics.label_font_size,
-    h_policy = Geyser.Fixed
-  }, __PKGNAME__.HPContainer)
-  __PKGNAME__.HPLabel:echo(nil, "nocolor", nil)
-
-  __PKGNAME__.HPBar = __PKGNAME__.HPBar or Geyser.Gauge:new({
-    name = "HPBar",
-  }, __PKGNAME__.HPContainer)
-  __PKGNAME__.HPBar:setStyleSheet(
-    __PKGNAME__.styles.HPFront,
-    __PKGNAME__.styles.HPBack,
-    __PKGNAME__.styles.GaugeText
-  )
-  __PKGNAME__.HPBar.text:setFont(__PKGNAME__.styles.WidgetFontName)
-  __PKGNAME__.HPBar.text:setFontSize(__PKGNAME__.metrics.gaugeFontSize)
-  __PKGNAME__.HPBar.text:echo(nil, "nocolor", nil)
-
-  -- SP
-  __PKGNAME__.SPContainer = __PKGNAME__.SPContainer or Geyser.HBox:new({
-    name = "SPContainer",
-  }, __PKGNAME__.VitalsBox);
-
-  __PKGNAME__.SPLabel = __PKGNAME__.SPLabel or Geyser.Label:new({
-    message = "SP",
-    width = 30,
-    name = "SPLabel",
-    stylesheet = __PKGNAME__.styles.Label,
-    fontSize = __PKGNAME__.metrics.label_font_size,
-    h_policy = Geyser.Fixed,
-  }, __PKGNAME__.SPContainer)
-  __PKGNAME__.SPLabel:echo(nil, "nocolor", nil)
-
-  __PKGNAME__.SPBar = __PKGNAME__.SPBar or Geyser.Gauge:new({
-    name = "SPBar",
-  }, __PKGNAME__.SPContainer)
-  __PKGNAME__.SPBar:setStyleSheet(
-    __PKGNAME__.styles.SPFront,
-    __PKGNAME__.styles.SPBack,
-    __PKGNAME__.styles.GaugeText
-  )
-  __PKGNAME__.SPBar.text:setFontSize(__PKGNAME__.metrics.gaugeFontSize)
-  __PKGNAME__.SPBar.text:echo(nil, "nocolor", nil)
-
-  -- MP
-  __PKGNAME__.MPContainer = __PKGNAME__.MPContainer or Geyser.HBox:new({
-    name = "MPContainer",
-  }, __PKGNAME__.VitalsBox);
-
-  __PKGNAME__.MPLabel = __PKGNAME__.MPLabel or Geyser.Label:new({
-    name = "MPLabel",
-    width = 30,
-    message = "MP",
-    stylesheet = __PKGNAME__.styles.Label,
-    fontSize = __PKGNAME__.metrics.label_font_size,
-    h_policy = Geyser.Fixed
-  }, __PKGNAME__.MPContainer)
-  __PKGNAME__.MPLabel:echo(nil, "nocolor", nil)
-
-  __PKGNAME__.MPBar = __PKGNAME__.MPBar or Geyser.Gauge:new({
-    name = "MPBar",
-  }, __PKGNAME__.MPContainer)
-  __PKGNAME__.MPBar:setStyleSheet(
-    __PKGNAME__.styles.MPFront,
-    __PKGNAME__.styles.MPBack,
-    __PKGNAME__.styles.GaugeText
-  )
-  __PKGNAME__.MPBar.text:setFontSize(__PKGNAME__.metrics.gaugeFontSize)
-  __PKGNAME__.MPBar.text:echo(nil, "nocolor", nil)
-
-  function __PKGNAME__.UpdateVitals()
-    if gmcp.Char.Vitals.hp ~= nil then __PKGNAME__.vitals.HP.current = tonumber(gmcp.Char.Vitals.hp) end
-    if gmcp.Char.Vitals.max_hp ~= nil then __PKGNAME__.vitals.HP.max = tonumber(gmcp.Char.Vitals.max_hp) end
-    if gmcp.Char.Vitals.sp ~= nil then __PKGNAME__.vitals.SP.current = tonumber(gmcp.Char.Vitals.sp) end
-    if gmcp.Char.Vitals.max_sp ~= nil then __PKGNAME__.vitals.SP.max = tonumber(gmcp.Char.Vitals.max_sp) end
-    if gmcp.Char.Vitals.mp ~= nil then __PKGNAME__.vitals.MP.current = tonumber(gmcp.Char.Vitals.mp) end
-    if gmcp.Char.Vitals.max_mp ~= nil then __PKGNAME__.vitals.MP.max = tonumber(gmcp.Char.Vitals.max_mp) end
-
-    if __PKGNAME__.vitals and __PKGNAME__.vitals.HP and __PKGNAME__.vitals.HP.current and __PKGNAME__.vitals.HP.max then
-      __PKGNAME__.UpdateBar(
-        __PKGNAME__.HPBar,
-        __PKGNAME__.vitals.HP.current,
-        __PKGNAME__.vitals.HP.max
-      )
-    end
-
-    if __PKGNAME__.vitals and __PKGNAME__.vitals.SP and __PKGNAME__.vitals.SP.current and __PKGNAME__.vitals.SP.max then
-      __PKGNAME__.UpdateBar(
-        __PKGNAME__.SPBar,
-        __PKGNAME__.vitals.SP.current,
-        __PKGNAME__.vitals.SP.max
-      )
-    end
-
-    if __PKGNAME__.vitals and __PKGNAME__.vitals.MP and __PKGNAME__.vitals.MP.current and __PKGNAME__.vitals.MP.max then
-      __PKGNAME__.UpdateBar(
-        __PKGNAME__.MPBar,
-        __PKGNAME__.vitals.MP.current,
-        __PKGNAME__.vitals.MP.max
-      )
-    end
-  end
-
-  -- Foe, XP, etc
-  __PKGNAME__.OtherBox = __PKGNAME__.OtherBox or Geyser.VBox:new({
-    name = "OtherBox",
-    width = "45%",
-    h_policy = Geyser.Fixed
-  }, __PKGNAME__.BarBox)
-
-  -- Foe
-  __PKGNAME__.FoeContainer = __PKGNAME__.FoeContainer or Geyser.HBox:new({
-    name = "FoeContainer",
-  }, __PKGNAME__.OtherBox);
-
-  __PKGNAME__.FoeLabel = __PKGNAME__.FoeLabel or Geyser.Label:new({
-    name = "FoeLabel",
-    width = 60,
-    message = "Foe",
-    stylesheet = __PKGNAME__.styles.Label,
-    fontSize = __PKGNAME__.metrics.label_font_size,
-    h_policy = Geyser.Fixed
-  }, __PKGNAME__.FoeContainer)
-  __PKGNAME__.FoeLabel:echo(nil, "nocolor", nil)
-
-  __PKGNAME__.FoeBar = Geyser.Gauge:new({
-    name = "FoeBar",
-  }, __PKGNAME__.FoeContainer)
-  __PKGNAME__.FoeBar:setStyleSheet(
-    __PKGNAME__.styles.FoeFront,
-    __PKGNAME__.styles.FoeBack,
-    __PKGNAME__.styles.GaugeText
-  )
-  __PKGNAME__.FoeBar.text:setFontSize(__PKGNAME__.metrics.gaugeFontSize)
-  __PKGNAME__.FoeBar.text:echo(nil, "nocolor", nil)
-  __PKGNAME__.UpdateBar(__PKGNAME__.FoeBar, 0, 100, "None")
-
-  -- XP
-  __PKGNAME__.XPContainer = __PKGNAME__.XPContainer or Geyser.HBox:new({
-    name = "XPContainer",
-  }, __PKGNAME__.OtherBox);
-
-  __PKGNAME__.XPLabel = __PKGNAME__.XPLabel or Geyser.Label:new({
-    name = "XPLabel",
-    width = 60,
-    message = "XP",
-    stylesheet = __PKGNAME__.styles.Label,
-    fontSize = __PKGNAME__.metrics.label_font_size,
-    h_policy = Geyser.Fixed
-  }, __PKGNAME__.XPContainer)
-  __PKGNAME__.XPLabel:echo(nil, "nocolor", nil)
-
-  __PKGNAME__.XPBar = Geyser.Gauge:new({
-    name = "XPBar",
-  }, __PKGNAME__.XPContainer)
-  __PKGNAME__.XPBar:setStyleSheet(
-    __PKGNAME__.styles.XPFront,
-    __PKGNAME__.styles.XPBack,
-    __PKGNAME__.styles.GaugeText
-  )
-  __PKGNAME__.XPBar.text:setFontSize(__PKGNAME__.metrics.gaugeFontSize)
-  __PKGNAME__.XPBar.text:echo(nil, "nocolor", nil)
-
-  __PKGNAME__.Status = __PKGNAME__.Status or {}
-  __PKGNAME__.Status.Advancement = __PKGNAME__.Status.Advancement or {}
-
-  -- Capacity
-  __PKGNAME__.CapBox = __PKGNAME__.CapBox or Geyser.HBox:new({
-    name = "CapBox",
-  }, __PKGNAME__.OtherBox)
-
-  -- Capacity
-  __PKGNAME__.CapContainer = __PKGNAME__.CapContainer or Geyser.HBox:new({
-    name = "CapContainer",
-    x = 0,
-    y = 0,
-    height = "100%",
-    width = "100%",
-  }, __PKGNAME__.CapBox);
-
-  __PKGNAME__.CapLabel = __PKGNAME__.CapLabel or Geyser.Label:new({
-    name = "CapLabel",
-    x = 0,
-    y = 0,
-    height = "100%",
-    width = 60,
-    message = "Capacity",
-    stylesheet = __PKGNAME__.styles.Label,
-    fontSize = __PKGNAME__.metrics.label_font_size,
-    h_policy = Geyser.Fixed,
-  }, __PKGNAME__.CapContainer)
-  __PKGNAME__.CapLabel:echo(nil, "nocolor", nil)
-
-  __PKGNAME__.CapBar = Geyser.Gauge:new({
-    name = "CapBar",
-    y = "20%",
-    height = "75%",
-  }, __PKGNAME__.CapContainer)
-  __PKGNAME__.CapBar:setStyleSheet(
-    __PKGNAME__.styles.CapFront,
-    __PKGNAME__.styles.CapBack,
-    __PKGNAME__.styles.GaugeText
-  )
-  __PKGNAME__.CapBar.text:setFontSize(__PKGNAME__.metrics.gaugeFontSize)
-  __PKGNAME__.CapBar.text:echo(nil, "nocolor", nil)
-
-  function __PKGNAME__.UpdateXP()
-    if not gmcp.Char.Status.xp or not gmcp.Char.Status.tnl then
-      return
-    end
-
-    local xp = tonumber(gmcp.Char.Status.xp)
-    local tnl = tonumber(gmcp.Char.Status.tnl)
-    local per = math.floor((xp / tnl) * 100)
-
-    __PKGNAME__.Status.Advancement = {
-      xp = xp,
-      tnl = tnl,
-      per = per
-    }
-
-    __PKGNAME__.UpdateBar(__PKGNAME__.XPBar, xp, tnl)
-  end
-
-  function __PKGNAME__.UpdateFoe()
-    if not gmcp.Char.Status.current_enemy then
-      return
-    end
-
-    local enemy = gmcp.Char.Status.current_enemy
-    local enemy_health
-
-    if enemy == nil or enemy == "" then
-      enemy = "None"
-      enemy_health = 0
-    else
-      enemy_health = tonumber(gmcp.Char.Status.current_enemy_health)
-    end
-
-    __PKGNAME__.UpdateBar(__PKGNAME__.FoeBar, enemy_health, 100, enemy)
-  end
-
-  function __PKGNAME__.UpdateCapacity()
-    if not gmcp.Char.Status.fill or not gmcp.Char.Status.capacity then
-      return
-    end
-
-    local fill = tonumber(gmcp.Char.Status.fill)
-    local cap = tonumber(gmcp.Char.Status.capacity)
-
-    __PKGNAME__.UpdateBar(__PKGNAME__.CapBar, fill, cap);
-  end
-
-  local handler
-  printDebug("", true)
-  handler = __PKGNAME__.config.package_name .. ":UpdateVitals"
-  if registerNamedEventHandler(
-        __PKGNAME__.config.package_name, handler, "gmcp.Char.Vitals", "__PKGNAME__:UpdateVitals"
-      ) then
-    __PKGNAME__.event_handlers[#__PKGNAME__.event_handlers + 1] = handler
-  end
-
-  handler = __PKGNAME__.config.package_name .. ":UpdateXP"
-  if registerNamedEventHandler(
-        __PKGNAME__.config.package_name, handler, "gmcp.Char.Status", "__PKGNAME__:UpdateXP"
-      ) then
-    __PKGNAME__.event_handlers[#__PKGNAME__.event_handlers + 1] = handler
-  end
-
-  handler = __PKGNAME__.config.package_name .. ":UpdateFoe"
-  if registerNamedEventHandler(
-        __PKGNAME__.config.package_name, handler, "gmcp.Char.Status", "__PKGNAME__:UpdateFoe"
-      ) then
-    __PKGNAME__.event_handlers[#__PKGNAME__.event_handlers + 1] = handler
-  end
-
-  handler = __PKGNAME__.config.package_name .. ":UpdateCapacity"
-  if registerNamedEventHandler(
-        __PKGNAME__.config.package_name, handler, "gmcp.Char.Status", "__PKGNAME__:UpdateCapacity"
-      ) then
-    __PKGNAME__.event_handlers[#__PKGNAME__.event_handlers + 1] = handler
-  end
+  __PKGNAME__.InventoryList = __PKGNAME__.InventoryList or {}
 
   __PKGNAME__.PanelWindow = __PKGNAME__.PanelWindow or Geyser.UserWindow:new({
     name = "PanelWindow",
@@ -332,8 +12,8 @@ function __PKGNAME__.buildUi()
     height = "100%",
     titleText = "__PKGNAME__",
     docked = true,
-    dockPosition = "l",
-    restoreLayout = true,
+    dockPosition = "r",
+    restoreLayout = false,
   })
 
   __PKGNAME__.Panel = __PKGNAME__.Panel or Geyser.Label:new({
@@ -392,16 +72,16 @@ function __PKGNAME__.buildUi()
   __PKGNAME__.InventoryRoom:setFontSize(__PKGNAME__.metrics.inventory_font_size)
   __PKGNAME__.InventoryRoom:disableScrolling()
 
-  -- -- Splitter
-  -- __PKGNAME__.SplitterLabel1 = __PKGNAME__.SplitterLabel1 or Geyser.Label:new({
-  --   name = "SplitterLabel1",
-  --   height = __PKGNAME__.splitter.size,
-  --   v_policy = Geyser.Fixed,
-  -- }, __PKGNAME__.InventoryContainer)
-  -- __PKGNAME__.SplitterLabel1:setFontSize(__PKGNAME__.splitter.font_size)
-  -- __PKGNAME__.SplitterLabel1:echo(nil, "nocolor", nil)
-  -- __PKGNAME__.SplitterLabel1:setStyleSheet(__PKGNAME__.styles.SplitterLabel)
-  -- __PKGNAME__.SplitterLabel1:echo(__PKGNAME__.splitter.symbol)
+  -- Splitter
+  __PKGNAME__.SplitterLabel1 = __PKGNAME__.SplitterLabel1 or Geyser.Label:new({
+    name = "SplitterLabel1",
+    height = __PKGNAME__.splitter.size,
+    v_policy = Geyser.Fixed,
+  }, __PKGNAME__.InventoryContainer)
+  __PKGNAME__.SplitterLabel1:setFontSize(__PKGNAME__.splitter.font_size)
+  __PKGNAME__.SplitterLabel1:echo(nil, "nocolor", nil)
+  __PKGNAME__.SplitterLabel1:setStyleSheet(__PKGNAME__.styles.SplitterLabel)
+  __PKGNAME__.SplitterLabel1:echo(__PKGNAME__.splitter.symbol)
 
   __PKGNAME__.InventoryInvContainer = __PKGNAME__.InventoryInvContainer or Geyser.VBox:new({
     name = "InventoryInvContainer",
@@ -427,6 +107,13 @@ function __PKGNAME__.buildUi()
   __PKGNAME__.InventoryInv:setFontSize(__PKGNAME__.metrics.inventory_font_size)
   __PKGNAME__.InventoryInv:disableScrolling()
 
+  -- Keep the section header borders (styles.Panel) above their Lists. Each List
+  -- sits directly below its header and, created later, is on top in z-order, so
+  -- as the VBox re-rounds positions on resize the List's top edge can clip the
+  -- header's 1px bottom border. Raising the headers lets their border always win.
+  raiseWindow(__PKGNAME__.InventoryRoomLabel.name)
+  raiseWindow(__PKGNAME__.InventoryInvLabel.name)
+
   __PKGNAME__.CoinLabel = __PKGNAME__.CoinLabel or Geyser.Label:new({
     name = "CoinLabel",
     height = 25,
@@ -434,24 +121,25 @@ function __PKGNAME__.buildUi()
     stylesheet = __PKGNAME__.styles.MainBG,
   }, __PKGNAME__.InventoryInvContainer)
 
-  -- __PKGNAME__.Splitter1 = __PKGNAME__.Splitter1 or Splitter:new({
-  --   name = "__PKGNAME__:Splitter1",
-  --   orientation = "vertical",
-  --   top = __PKGNAME__.InventoryRoomContainer,
-  --   middle = __PKGNAME__.SplitterLabel1,
-  --   bottom = __PKGNAME__.InventoryInvContainer,
-  --   margins = {
-  --     top = {
-  --       __PKGNAME__.InventoryRoomLabel.get_height(),
-  --       0,
-  --     },
-  --     bottom = {
-  --       0,
-  --       __PKGNAME__.InventoryInvLabel.get_height() +
-  --       __PKGNAME__.CoinLabel.get_height(),
-  --     }
-  --   }
-  -- })
+  __PKGNAME__.Splitter1 = __PKGNAME__.Splitter1 or Geyser.Splitter:new({
+    name = "__PKGNAME__:Splitter1",
+    orientation = "vertical",
+    anchor = "bottom",
+    top = __PKGNAME__.InventoryRoomContainer,
+    middle = __PKGNAME__.SplitterLabel1,
+    bottom = __PKGNAME__.InventoryInvContainer,
+    margins = {
+      top = {
+        __PKGNAME__.InventoryRoomLabel.get_height(),
+        0,
+      },
+      bottom = {
+        0,
+        __PKGNAME__.InventoryInvLabel.get_height() +
+        __PKGNAME__.CoinLabel.get_height(),
+      }
+    }
+  })
 
   __PKGNAME__.CoinBox = __PKGNAME__.CoinBox or Geyser.HBox:new({
     x = 0,
@@ -471,15 +159,15 @@ function __PKGNAME__.buildUi()
     width = 25,
     h_policy = Geyser.Fixed,
     stylesheet = __PKGNAME__.styles.CoinLabel,
-    message = __PKGNAME__.coin.symbol,
-    fontSize = __PKGNAME__.metrics.coin_font_size,
+    message = "",
   }, __PKGNAME__.PlatinumBox)
-  __PKGNAME__.PlatinumLabel:echo(nil, "nocolor", nil)
-  __PKGNAME__.PlatinumLabel:setStyleSheet(f "{__PKGNAME__.styles.Center} {__PKGNAME__.styles.CoinPlatinum}")
+  __PKGNAME__.PlatinumLabel:setBackgroundImage(
+    __PKGNAME__.config.package_path .. "coins/platinum.png"
+  )
 
   __PKGNAME__.PlatinumNumber = __PKGNAME__.PlatinumNumber or Geyser.Label:new({
     name = "PlatinumNumber",
-    stylesheet = __PKGNAME__.styles.CoinLabel,
+    stylesheet = __PKGNAME__.styles.CoinNumber,
     message = "0",
   }, __PKGNAME__.PlatinumBox)
 
@@ -493,21 +181,21 @@ function __PKGNAME__.buildUi()
     width = 25,
     h_policy = Geyser.Fixed,
     stylesheet = __PKGNAME__.styles.CoinLabel,
-    message = __PKGNAME__.coin.symbol,
-    fontSize = __PKGNAME__.metrics.coin_font_size,
+    message = "",
   }, __PKGNAME__.GoldBox)
-
-  __PKGNAME__.GoldLabel:echo(nil, "nocolor", nil)
-  __PKGNAME__.GoldLabel:setStyleSheet(f "{__PKGNAME__.styles.Center} {__PKGNAME__.styles.CoinGold}")
+  __PKGNAME__.GoldLabel:setBackgroundImage(
+    __PKGNAME__.config.package_path .. "coins/gold.png"
+  )
 
   __PKGNAME__.GoldNumber = __PKGNAME__.GoldNumber or Geyser.Label:new({
     name = "GoldNumber",
-    stylesheet = __PKGNAME__.styles.CoinLabel,
+    stylesheet = __PKGNAME__.styles.CoinNumber,
     message = "0",
   }, __PKGNAME__.GoldBox)
 
   __PKGNAME__.SilverBox = __PKGNAME__.SilverBox or Geyser.HBox:new({
     name = "SilverBox",
+    height = "100%",
   }, __PKGNAME__.CoinBox)
 
   __PKGNAME__.SilverLabel = __PKGNAME__.SilverLabel or Geyser.Label:new({
@@ -515,16 +203,15 @@ function __PKGNAME__.buildUi()
     width = 25,
     h_policy = Geyser.Fixed,
     stylesheet = __PKGNAME__.styles.CoinLabel,
-    message = __PKGNAME__.coin.symbol,
-    fontSize = __PKGNAME__.metrics.coin_font_size,
+    message = "",
   }, __PKGNAME__.SilverBox)
-
-  __PKGNAME__.SilverLabel:echo(nil, "nocolor", nil)
-  __PKGNAME__.SilverLabel:setStyleSheet(f "{__PKGNAME__.styles.Center} {__PKGNAME__.styles.CoinSilver}")
+  __PKGNAME__.SilverLabel:setBackgroundImage(
+    __PKGNAME__.config.package_path .. "coins/silver.png"
+  )
 
   __PKGNAME__.SilverNumber = __PKGNAME__.SilverNumber or Geyser.Label:new({
     name = "SilverNumber",
-    stylesheet = __PKGNAME__.styles.CoinLabel,
+    stylesheet = __PKGNAME__.styles.CoinNumber,
     message = "0",
   }, __PKGNAME__.SilverBox)
 
@@ -538,16 +225,15 @@ function __PKGNAME__.buildUi()
     width = 25,
     h_policy = Geyser.Fixed,
     stylesheet = __PKGNAME__.styles.CoinLabel,
-    message = __PKGNAME__.coin.symbol,
-    fontSize = __PKGNAME__.metrics.coin_font_size,
+    message = "",
   }, __PKGNAME__.CopperBox)
-
-  __PKGNAME__.CopperLabel:echo(nil, "nocolor", nil)
-  __PKGNAME__.CopperLabel:setStyleSheet(f "{__PKGNAME__.styles.Center} {__PKGNAME__.styles.CoinCopper}")
+  __PKGNAME__.CopperLabel:setBackgroundImage(
+    __PKGNAME__.config.package_path .. "coins/copper.png"
+  )
 
   __PKGNAME__.CopperNumber = __PKGNAME__.CopperNumber or Geyser.Label:new({
     name = "CopperNumber",
-    stylesheet = __PKGNAME__.styles.CoinLabel,
+    stylesheet = __PKGNAME__.styles.CoinNumber,
     message = "0",
   }, __PKGNAME__.CopperBox)
 
@@ -642,6 +328,8 @@ function __PKGNAME__.buildUi()
     end
   end
 
+  local handler
+
   handler = __PKGNAME__.config.package_name .. ":UpdateCoins"
   if registerNamedEventHandler(
         __PKGNAME__.config.package_name, handler, "gmcp.Char.Status", "__PKGNAME__:UpdateCoins"
@@ -703,20 +391,23 @@ function __PKGNAME__.buildUi()
     -- Depending on which inventory we're looking at, we'll need to update
     -- the appropriate table and widget
     local location, widget, table_name = __PKGNAME__.DetermineUpdateInfo(gmcp.Char.Items.List)
+    if not table_name then return end
     __PKGNAME__[table_name] = table.deepcopy(gmcp.Char.Items.List.items)
-    __PKGNAME__.UpdateInventoryWidget(location, widget, __PKGNAME__[table_name])
+    __PKGNAME__.ScheduleInventoryUpdate(location, widget, table_name)
   end
 
   function __PKGNAME__.AddInventory(event, ...)
     if not __PKGNAME__ then return end
 
     local location, widget, table_name = __PKGNAME__.DetermineUpdateInfo(gmcp.Char.Items.Add)
+    if not table_name then return end
     table.insert(__PKGNAME__[table_name], 1, table.deepcopy(gmcp.Char.Items.Add.item))
-    __PKGNAME__.UpdateInventoryWidget(location, widget, __PKGNAME__[table_name])
+    __PKGNAME__.ScheduleInventoryUpdate(location, widget, table_name)
   end
 
   function __PKGNAME__.RemoveInventory(event, ...)
     local location, widget, table_name = __PKGNAME__.DetermineUpdateInfo(gmcp.Char.Items.Remove)
+    if not table_name then return end
 
     for i, item in pairs(__PKGNAME__[table_name]) do
       if item.hash == gmcp.Char.Items.Remove.item.hash then
@@ -725,11 +416,12 @@ function __PKGNAME__.buildUi()
       end
     end
 
-    __PKGNAME__.UpdateInventoryWidget(location, widget, __PKGNAME__[table_name])
+    __PKGNAME__.ScheduleInventoryUpdate(location, widget, table_name)
   end
 
   function __PKGNAME__.UpdateInventory(event, ...)
     local location, widget, table_name = __PKGNAME__.DetermineUpdateInfo(gmcp.Char.Items.Update)
+    if not table_name then return end
     for i, item in pairs(__PKGNAME__[table_name]) do
       if item.hash == gmcp.Char.Items.Update.item.hash then
         __PKGNAME__[table_name][i] = table.deepcopy(gmcp.Char.Items.Update.item)
@@ -737,7 +429,28 @@ function __PKGNAME__.buildUi()
       end
     end
 
-    __PKGNAME__.UpdateInventoryWidget(location, widget, __PKGNAME__[table_name])
+    __PKGNAME__.ScheduleInventoryUpdate(location, widget, table_name)
+  end
+
+  __PKGNAME__.update_timers = __PKGNAME__.update_timers or {}
+
+  -- Coalesce bursts of GMCP item events (e.g. "drop all" / "get all", which fire
+  -- a flurry of Char.Items.Add/Remove) into a single widget rebuild. Every event
+  -- still mutates the data table immediately; only the expensive row rebuild is
+  -- debounced -- without this, dropping N items triggers N full rebuilds (O(n^2)
+  -- Label churn). Keyed by location so room and inv debounce independently. The
+  -- table is re-read at fire time (not captured) since ListInventory replaces it.
+  function __PKGNAME__.ScheduleInventoryUpdate(location, widget, table_name)
+    local pending = __PKGNAME__.update_timers[location]
+    if pending then
+      killTimer(pending)
+    end
+
+    __PKGNAME__.update_timers[location] = tempTimer(__PKGNAME__.metrics.inventory_debounce, function()
+      __PKGNAME__.update_timers[location] = nil
+      if not __PKGNAME__ then return end
+      __PKGNAME__.UpdateInventoryWidget(location, widget, __PKGNAME__[table_name])
+    end)
   end
 
   function __PKGNAME__.UpdateInventoryWidget(location, widget, inventory)
@@ -774,8 +487,8 @@ function __PKGNAME__.buildUi()
   end
 
   function __PKGNAME__.Disconnect()
-    __PKGNAME__.inventory.room = {}
-    __PKGNAME__.inventory.inv = {}
+    __PKGNAME__.RoomInventoryList = {}
+    __PKGNAME__.InventoryList = {}
 
     __PKGNAME__.UpdateInventoryWidget("room", __PKGNAME__.InventoryRoom, __PKGNAME__.RoomInventoryList)
     __PKGNAME__.UpdateInventoryWidget("inv", __PKGNAME__.InventoryInv, __PKGNAME__.InventoryList)
@@ -814,5 +527,15 @@ function __PKGNAME__.buildUi()
         __PKGNAME__.config.package_name, handler, "sysDisconnectionEvent", "__PKGNAME__:Disconnect"
       ) then
     __PKGNAME__.event_handlers[#__PKGNAME__.event_handlers + 1] = handler
+  end
+end
+
+function __PKGNAME__.TeardownPanelWindow()
+  if __PKGNAME__.Splitter1 then
+    __PKGNAME__.Splitter1:disconnect()
+  end
+
+  if __PKGNAME__.PanelWindow then
+    __PKGNAME__.PanelWindow:delete()
   end
 end
